pyproject.toml
[build-system]
requires = ["setuptools>=80.9.0", "wheel>=0.45.1"]
build-backend = "setuptools.build_meta"

[project]
name = "alloy"
version = "0.1.0"
description = "Alloy (Python): Python for logic. English for intelligence."
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "George Lydakis" }]
license = { text = "MIT" }
dependencies = [
  "openai>=1.99.6,<2",
  "python-dotenv>=1.1.1,<2",
]
classifiers = [
  "Intended Audience :: Developers",
  "Operating System :: OS Independent",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
]

[project.urls]
Homepage = "https://github.com/openai/alloy-py"
Repository = "https://github.com/openai/alloy-py"
Issues = "https://github.com/openai/alloy-py/issues"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]
include = ["alloy*"]

[tool.setuptools.package-data]
alloy = ["py.typed", "__init__.pyi"]

[project.optional-dependencies]
openai = ["openai>=1.99.6,<2", "python-dotenv>=1.1.1,<2"]
anthropic = ["anthropic>=0.62.0,<0.63"]
# New GA SDK
gemini = ["google-genai>=1.29.0,<2"]
# Legacy SDK
gemini-legacy = ["google-generativeai>=0.8.5,<0.9"]
providers = [
  "openai>=1.99.6,<2",
  "python-dotenv>=1.1.1,<2",
  "anthropic>=0.62.0,<0.63",
  "google-genai>=1.29.0,<2",
  "ollama>=0.5.3,<0.6",
]
ollama = ["ollama>=0.5.3,<0.6"]
dev = [
  "pytest>=8.4.1,<9",
  "pytest-asyncio>=0.25.2,<0.26",
  "ruff>=0.12.8,<0.13",
  "black>=25.1.0,<26",
  "mypy>=1.17.1,<2",
  "pre-commit>=4.3.0,<5",
]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.mypy]
python_version = "3.10"
warn_unused_ignores = true
ignore_missing_imports = true
strict_optional = true
show_error_codes = true
pretty = true
exclude = [".venv/"]
